AccessControl.ino: findID omits last entry
I've been playing around with the AccessControl example all day and noticed that sometimes when I tried to delete a known tag it would return "don't know this id" instead and add the tag again, when in fact, it was already written in the eeprom (locally, on my machine, I printed out the entirety of the EEPROM and it was in fact listed).

I noticed that this was only the case on the last added tag and I believe that in the findID() the loop (since it starts at 1) needs to reach count, i.e. i<= count and not i<count (like in findIDSLOT).